Set px to pre tag
If I set %  to the height of the pre tag, the tag cannot display scroll when the chat log' height get longer than the pre tag's one. So, for now I set 500px. However, if I set pixel the appearance of height change depending on  each resolution of your device. I want to fix it.